refactoted giant validator file
